Major update. Reworked the Order interface and how the Depot, Portfolio and Positions are handeled.
The Depot is now only used to keep track of the cash. The Portfolio
is only used to manage which Positions are tracked and the
Positions only track the number of shares and the prices.
Each of these objects now has their own History object.
Added base_positions to the Portfolio. Orders now have to target an
existing Position or a CandleFeed. If a CandleFeed is targeted the
base_position in the according Portfolio will be used to track the
changes.
Added an OrderStatus and OrderAction class to keep track of what
the Orders do. The Broker now uses a new object called
FillingStrategy to calculate how the order is filled.
Next on the roadmap are the implementation of strategies.