Changed a core feature of DateSeries. They are now using Events.
Events can be used to synchronize multiple DateSeries. All
actions on the data and head of a DateSeries will emit a signal
that is listened for by all DateSeries that are children of the
emitting DateSeries. (In principle all DateSeries assigned to the
same EventHandler listen to the signal. However, by default only
children act upon receiving such a signal.)

This commit aims to re-code all of the DateSeries and math
functionality of before. Only the Crossover class from the
math.indicator module is lost and will be added back in at a later
stage.